+++ pochidetection/models/ssdlite.py
@@ -7,7 +7,7 @@ from torchvision.models import MobileNet_V3_Large_Weights
 from torchvision.models.detection import ssdlite320_mobilenet_v3_large
 from torchvision.models.detection.ssd import SSD
 
-from pochidetection.interfaces.model import IDetectionModel, ModelOutputDict
+from pochidetection.interfaces.model import IDetectionModel, SSDModelOutputDict
 
 
 class SSDLiteModel(IDetectionModel):
@@ -63,7 +63,7 @@ class SSDLiteModel(IDetectionModel):
         self,
         pixel_values: torch.Tensor,
         labels: list[dict[str, torch.Tensor]] | None = None,
-    ) -> ModelOutputDict:
+    ) -> SSDModelOutputDict:
         """順伝播.
 
         Args: